+++ campaign-loyalty-system/backend/src/main/java/com/campaignloyalty/qrtoken/repository/QrTokenRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface QrTokenRepository extends JpaRepository<QrToken, Long> {
+public interface QrTokenRepository extends JpaRepository<QrToken, Integer> {
 
     QrToken findByToken(String token);
