+++ campaign-loyalty-system/backend/src/main/java/com/campaignloyalty/reward/entity/Reward.java
@@ -12,13 +12,13 @@ public class Reward {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "customer_id")
-    private Long customerId;
+    private Integer customerId;
 
     @Column(name = "hotel_id")
-    private Long hotelId;
+    private Integer hotelId;
 
     @Column(name = "earned_at")
     private LocalDateTime earnedAt;
