+++ campaign-loyalty-system/backend/src/main/java/com/campaignloyalty/reward/repository/RewardRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDateTime;
 
-public interface RewardRepository extends JpaRepository<Reward, Long> {
+public interface RewardRepository extends JpaRepository<Reward, Integer> {
 
-    long countByHotelIdAndEarnedAtBetween(Long hotelId, LocalDateTime start, LocalDateTime end);
+    long countByHotelIdAndEarnedAtBetween(Integer hotelId, LocalDateTime start, LocalDateTime end);
 
     long countByEarnedAtBetween(LocalDateTime start, LocalDateTime end);
 }
