+++ campaign-loyalty-system/backend/src/main/java/com/campaignloyalty/scan/entity/ScanHistory.java
@@ -12,13 +12,13 @@ public class ScanHistory {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "customer_id")
-    private Long customerId;
+    private Integer customerId;
 
     @Column(name = "hotel_id")
-    private Long hotelId;
+    private Integer hotelId;
 
     @Column(name = "qr_token")
     private String qrToken;
